pyproject.toml
[project]
name = "dinotool"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Mikko Impiö", email = "mikko.impio@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "einops>=0.8.1",
    "numpy>=2.2.4",
    "opencv-python>=4.11.0.86",
    "pillow>=11.1.0",
    "scikit-learn>=1.6.1",
    "torch>=2.6.0",
    "torchvision>=0.21.0",
    "tqdm>=4.67.1",
    "xformers>=0.0.29.post3",
]

[project.scripts]
dinotool = "dinotool.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
    "ipywidgets>=8.1.5",
    "matplotlib>=3.10.1",
    "pytest>=8.3.5",
]
